treewide: break into multiple crates
