+++ Cargo.toml
@@ -1,36 +1,19 @@
-[package]
-name         = "microfetch"
-description  = "Microscopic fetch tool in Rust, for NixOS systems, with special emphasis on speed"
-version      = "0.4.13"
-edition      = "2024"
+[workspace]
+members  = [ "crates/*", "microfetch" ]
+resolver = "3"
+
+[workspace.package]
 authors      = [ "NotAShelf <raf@notashelf.dev>" ]
-rust-version = "1.91.0"
-readme       = "./docs/README.md"
-repository   = "https://github.com/notashelf/microfetch"
+edition      = "2024"
 license      = "GPL-3.0"
+rust-version = "1.92.0"
+version      = "0.4.13"
 
-[lib]
-name = "microfetch_lib"
-path = "src/lib.rs"
+[workspace.dependencies]
+microfetch-asm = { path = "./crates/asm" }
+microfetch-lib = { path = "./crates/lib" }
 
-[[bin]]
-name = "microfetch"
-path = "src/main.rs"
-
-[dependencies]
-hotpath = { optional = true, version = "0.13.0" }
-
-[dev-dependencies]
-criterion = "0.8.1"
-
-[features]
-hotpath       = [ "dep:hotpath", "hotpath/hotpath" ]
-hotpath-alloc = [ "hotpath/hotpath-alloc" ]
-hotpath-off   = [ "hotpath/hotpath-off" ]
-
-[[bench]]
-harness = false
-name    = "benchmark"
+criterion = { default-features = false, features = [ "cargo_bench_support" ], version = "0.8.2" }
 
 [profile.dev]
 opt-level = 1
@@ -48,15 +31,13 @@ inherits        = "release"
 split-debuginfo = "unpacked"
 strip           = "none"
 
-[lints.clippy]
+[workspace.lints.clippy]
 complexity = { level = "warn", priority = -1 }
 nursery    = { level = "warn", priority = -1 }
 pedantic   = { level = "warn", priority = -1 }
 perf       = { level = "warn", priority = -1 }
 style      = { level = "warn", priority = -1 }
 
-# The lint groups above enable some less-than-desirable rules, we should manually
-# enable those to keep our sanity.
 absolute_paths                 = "allow"
 arbitrary_source_item_ordering = "allow"
 implicit_return                = "allow"
