+++ crates/asm/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name                   = "microfetch-asm"
+description            = "Inline assembly syscall helpers for microfetch"
+version.workspace      = true
+edition.workspace      = true
+authors.workspace      = true
+rust-version.workspace = true
+publish                = false
+license.workspace      = true
+
+[lints]
+workspace = true
