+++ crates/lib/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name                   = "microfetch-lib"
+description            = "Microfetch library crate - exports all functionality including main function"
+version.workspace      = true
+edition.workspace      = true
+authors.workspace      = true
+rust-version.workspace = true
+license.workspace      = true
+publish                = false
+
+[dependencies]
+hotpath                  = { optional = true, version = "0.13.0" }
+microfetch-asm.workspace = true
+
+[features]
+hotpath       = [ "dep:hotpath", "hotpath/hotpath" ]
+hotpath-alloc = [ "hotpath/hotpath-alloc" ]
+hotpath-off   = [ "hotpath/hotpath-off" ]
+
+[dev-dependencies]
+criterion.workspace = true
+
+[[bench]]
+harness = false
+name    = "benchmark"
+
+[lints]
+workspace = true
