+++ crates/lib/src/lib.rs
@@ -1,41 +1,64 @@
-use std::io::{self, Cursor, Write};
+pub mod colors;
+pub mod desktop;
+pub mod release;
+pub mod system;
+pub mod uptime;
 
-use microfetch_lib::{
-  UtsName,
-  colors::{COLORS, print_dots},
-  desktop::get_desktop_info,
-  release::{get_os_pretty_name, get_system_info},
-  syscall::sys_write,
-  system::{
-    get_memory_usage,
-    get_root_disk_usage,
-    get_shell,
-    get_username_and_hostname,
-  },
-  uptime::get_current,
+use std::{
+  ffi::CStr,
+  io::{self, Cursor, Write},
+  mem::MaybeUninit,
 };
 
-#[cfg_attr(feature = "hotpath", hotpath::main)]
-fn main() -> Result<(), Box<dyn std::error::Error>> {
-  if Some("--version") == std::env::args().nth(1).as_deref() {
-    println!("Microfetch {}", env!("CARGO_PKG_VERSION"));
-  } else {
-    let utsname = UtsName::uname()?;
-    let fields = Fields {
-      user_info:      get_username_and_hostname(&utsname),
-      os_name:        get_os_pretty_name()?,
-      kernel_version: get_system_info(&utsname),
-      shell:          get_shell(),
-      desktop:        get_desktop_info(),
-      uptime:         get_current()?,
-      memory_usage:   get_memory_usage()?,
-      storage:        get_root_disk_usage()?,
-      colors:         print_dots(),
-    };
-    print_system_info(&fields)?;
+pub use microfetch_asm as syscall;
+pub use microfetch_asm::{
+  StatfsBuf,
+  UtsNameBuf,
+  read_file_fast,
+  sys_close,
+  sys_open,
+  sys_read,
+  sys_statfs,
+  sys_uname,
+  sys_write,
+};
+
+/// Wrapper for `utsname` with safe accessor methods
+pub struct UtsName(UtsNameBuf);
+
+impl UtsName {
+  /// Calls `uname(2)` syscall and returns a `UtsName` wrapper
+  ///
+  /// # Errors
+  ///
+  /// Returns an error if the `uname` syscall fails
+  pub fn uname() -> Result<Self, std::io::Error> {
+    let mut uts = MaybeUninit::uninit();
+    if unsafe { sys_uname(uts.as_mut_ptr()) } != 0 {
+      return Err(std::io::Error::last_os_error());
+    }
+    Ok(Self(unsafe { uts.assume_init() }))
   }
 
-  Ok(())
+  #[must_use]
+  pub const fn nodename(&self) -> &CStr {
+    unsafe { CStr::from_ptr(self.0.nodename.as_ptr().cast()) }
+  }
+
+  #[must_use]
+  pub const fn sysname(&self) -> &CStr {
+    unsafe { CStr::from_ptr(self.0.sysname.as_ptr().cast()) }
+  }
+
+  #[must_use]
+  pub const fn release(&self) -> &CStr {
+    unsafe { CStr::from_ptr(self.0.release.as_ptr().cast()) }
+  }
+
+  #[must_use]
+  pub const fn machine(&self) -> &CStr {
+    unsafe { CStr::from_ptr(self.0.machine.as_ptr().cast()) }
+  }
 }
 
 // Struct to hold all the fields we need in order to print the fetch. This
@@ -69,9 +92,9 @@ fn print_system_info(
     colors,
   } = fields;
 
-  let cyan = COLORS.cyan;
-  let blue = COLORS.blue;
-  let reset = COLORS.reset;
+  let cyan = colors::COLORS.cyan;
+  let blue = colors::COLORS.blue;
+  let reset = colors::COLORS.reset;
 
   let mut buf = [0u8; 2048];
   let mut cursor = Cursor::new(&mut buf[..]);
@@ -104,5 +127,34 @@ fn print_system_info(
         .into(),
     );
   }
+  Ok(())
+}
+
+/// Main entry point for microfetch - can be called by the binary crate
+/// or by other consumers of the library
+///
+/// # Errors
+///
+/// Returns an error if any system call fails
+#[cfg_attr(feature = "hotpath", hotpath::main)]
+pub fn run() -> Result<(), Box<dyn std::error::Error>> {
+  if Some("--version") == std::env::args().nth(1).as_deref() {
+    println!("Microfetch {}", env!("CARGO_PKG_VERSION"));
+  } else {
+    let utsname = UtsName::uname()?;
+    let fields = Fields {
+      user_info:      system::get_username_and_hostname(&utsname),
+      os_name:        release::get_os_pretty_name()?,
+      kernel_version: release::get_system_info(&utsname),
+      shell:          system::get_shell(),
+      desktop:        desktop::get_desktop_info(),
+      uptime:         uptime::get_current()?,
+      memory_usage:   system::get_memory_usage()?,
+      storage:        system::get_root_disk_usage()?,
+      colors:         colors::print_dots(),
+    };
+    print_system_info(&fields)?;
+  }
+
   Ok(())
 }
