+++ microfetch/src/main.rs
@@ -1,0 +1,4 @@
+#[cfg_attr(feature = "hotpath", hotpath::main)]
+fn main() -> Result<(), Box<dyn std::error::Error>> {
+  microfetch_lib::run()
+}
